skills/pexip-external-policy/examples/entra-avatar-lookup/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pillow>=10.0
python-dotenv>=1.0
